refactor: move shared computing unit types and services
