+++ frontend/src/app/common/type/computing-unit-connection.interface.ts
