+++ frontend/src/app/common/util/computing-unit.util.ts
@@ -19,7 +19,7 @@
 
 import { Component, inject } from "@angular/core";
 import { NZ_MODAL_DATA } from "ng-zorro-antd/modal";
-import { DashboardWorkflowComputingUnit } from "../../workspace/types/workflow-computing-unit";
+import { DashboardWorkflowComputingUnit } from "../type/workflow-computing-unit";
 
 @Component({
   template: `
