+++ frontend/src/app/dashboard/component/user/user-computing-unit/user-computing-unit-list-item/user-computing-unit-list-item.component.ts
@@ -27,7 +27,7 @@ import {
   ViewChild,
   ElementRef,
 } from "@angular/core";
-import { ComputingUnitStatusService } from "../../../../../workspace/service/computing-unit-status/computing-unit-status.service";
+import { ComputingUnitStatusService } from "../../../../../common/service/computing-unit/computing-unit-status/computing-unit-status.service";
 import { extractErrorMessage } from "../../../../../common/util/error";
 import { NotificationService } from "../../../../../common/service/notification/notification.service";
 import { NzModalService } from "ng-zorro-antd/modal";
@@ -35,8 +35,8 @@ import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import {
   DashboardWorkflowComputingUnit,
   WorkflowComputingUnit,
-} from "../../../../../workspace/types/workflow-computing-unit";
-import { WorkflowComputingUnitManagingService } from "../../../../../workspace/service/workflow-computing-unit/workflow-computing-unit-managing.service";
+} from "../../../../../common/type/workflow-computing-unit";
+import { WorkflowComputingUnitManagingService } from "../../../../../common/service/computing-unit/computing-unit-managing/workflow-computing-unit-managing.service";
 import {
   ComputingUnitMetadataComponent,
   parseResourceUnit,
@@ -53,7 +53,7 @@ import {
   getComputingUnitCpuLimitUnit,
 } from "../../../../../common/util/computing-unit.util";
 import { GuiConfigService } from "../../../../../common/service/gui-config.service";
-import { ComputingUnitActionsService } from "../../../../service/user/computing-unit-actions/computing-unit-actions.service";
+import { ComputingUnitActionsService } from "../../../../../common/service/computing-unit/computing-unit-actions/computing-unit-actions.service";
 
 @UntilDestroy()
 @Component({
