+++ frontend/src/app/workspace/component/property-editor/operator-property-edit-frame/operator-property-edit-frame.component.spec.ts
@@ -47,8 +47,8 @@ import { cloneDeep } from "lodash-es";
 import Ajv from "ajv";
 import { COLLAB_DEBOUNCE_TIME_MS } from "../../../../common/formly/collab-wrapper/collab-wrapper/collab-wrapper.component";
 import { FormlyNgZorroAntdModule } from "@ngx-formly/ng-zorro-antd";
-import { ComputingUnitStatusService } from "../../../service/computing-unit-status/computing-unit-status.service";
-import { MockComputingUnitStatusService } from "../../../service/computing-unit-status/mock-computing-unit-status.service";
+import { ComputingUnitStatusService } from "../../../../common/service/computing-unit/computing-unit-status/computing-unit-status.service";
+import { MockComputingUnitStatusService } from "../../../../common/service/computing-unit/computing-unit-status/mock-computing-unit-status.service";
 import { commonTestProviders } from "../../../../common/testing/test-utils";
 
 const { marbles } = configure({ run: false });
