+++ frontend/src/app/workspace/component/result-panel/error-frame/error-frame.component.spec.ts
@@ -24,8 +24,8 @@ import { OperatorMetadataService } from "../../../service/operator-metadata/oper
 import { StubOperatorMetadataService } from "../../../service/operator-metadata/stub-operator-metadata.service";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { NzDropDownModule } from "ng-zorro-antd/dropdown";
-import { ComputingUnitStatusService } from "src/app/workspace/service/computing-unit-status/computing-unit-status.service";
-import { MockComputingUnitStatusService } from "src/app/workspace/service/computing-unit-status/mock-computing-unit-status.service";
+import { ComputingUnitStatusService } from "../../../../common/service/computing-unit/computing-unit-status/computing-unit-status.service";
+import { MockComputingUnitStatusService } from "../../../../common/service/computing-unit/computing-unit-status/mock-computing-unit-status.service";
 import { commonTestProviders } from "../../../../common/testing/test-utils";
 
 describe("ConsoleFrameComponent", () => {
